# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Souklick Review Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get all reviews with optional filters
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
            enum: [google, zomato, tripadvisor]
        - name: rating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, draft_saved, responded, skipped]
        - name: priorityOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsPage"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review (for testing/demo purposes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a specific review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review details with responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewWithResponses"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateReviewStatus
      tags: [reviews]
      summary: Update review status (mark as skipped, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewStatusRequest"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /responses:
    post:
      operationId: createResponse
      tags: [responses]
      summary: Save a draft response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponseRequest"
      responses:
        "201":
          description: Created response draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /responses/{id}:
    patch:
      operationId: updateResponse
      tags: [responses]
      summary: Update a response (edit text, approve, reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResponseRequest"
      responses:
        "200":
          description: Updated response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /responses/{id}/approve:
    post:
      operationId: approveResponse
      tags: [responses]
      summary: Approve and post a response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveResponseRequest"
      responses:
        "200":
          description: Response posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Review:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        platform:
          type: string
          enum: [google, zomato, tripadvisor]
        platformReviewId:
          type: string
        reviewerName:
          type: string
        reviewerProfileUrl:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
          nullable: true
        reviewDate:
          type: string
          format: date-time
        responseStatus:
          type: string
          enum: [pending, draft_saved, responded, skipped]
        sentimentScore:
          type: number
          nullable: true
        locationName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - locationId
        - platform
        - platformReviewId
        - reviewerName
        - rating
        - reviewDate
        - responseStatus
        - locationName
        - createdAt
    ReviewWithResponses:
      allOf:
        - $ref: "#/components/schemas/Review"
        - type: object
          properties:
            responses:
              type: array
              items:
                $ref: "#/components/schemas/Response"
          required:
            - responses
    ReviewsPage:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - reviews
        - total
        - page
        - limit
        - totalPages
...
    CreateReviewRequest:
      type: object
      properties:
        locationId:
          type: string
        platform:
          type: string
          enum: [google, zomato, tripadvisor]
        reviewerName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
        reviewDate:
          type: string
          format: date-time
      required:
        - locationId
        - platform
        - reviewerName
        - rating
    UpdateReviewStatusRequest:
      type: object
      properties:
        responseStatus:
          type: string
          enum: [pending, draft_saved, responded, skipped]
      required:
        - responseStatus
    Response:
      type: object
      properties:
        id:
          type: string
        reviewId:
          type: string
        draftedBy:
          type: string
          nullable: true
        draftText:
          type: string
        finalText:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, approved, posted, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - reviewId
        - draftText
        - status
        - createdAt
        - updatedAt
    CreateResponseRequest:
      type: object
      properties:
        reviewId:
          type: string
        draftText:
          type: string
      required:
        - reviewId
        - draftText
    UpdateResponseRequest:
      type: object
      properties:
        draftText:
          type: string
        finalText:
          type: string
        status:
          type: string
          enum: [draft, approved, posted, failed]
    ApproveResponseRequest:
      type: object
      properties:
        finalText:
          type: string
      required:
        - finalText
...