# lib/api-zod/src/generated/types/reviewWithResponses.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Souklick Review Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Response } from "./response";
import type { Review } from "./review";

export type ReviewWithResponses = Review & {
  responses: Response[];
};